common: deduplicate ring_buffer with llama-impl.h
- Move ring_buffer template to src/llama-impl.h
- Remove duplicates from llama-sampler.cpp and sampling.cpp
- Merge improvements: capacity check, std::copy, better error messages
- Resolves TODO: deduplicate with llama-impl.h
